Changed the Mine class, so it can be stored more easily; Implemented WorldEditMine;

+++ src/main/java/com/shebbasoft/storm/mine/command/SetResetTimeCommand.java
@@ -37,9 +37,9 @@ public class SetResetTimeCommand extends SimpleCommand {
             return;
         }
 
-        int time;
+        long time;
         try {
-            time = Integer.parseInt(arguments[1]);
+            time = Long.parseLong(arguments[1]);
         } catch (NumberFormatException e) {
             sender.sendMessage(ChatColor.RED + "Error: Invalid time (in seconds)");
             return;
